Add json output option
Pair-programmed with @Vagabond

The 'json' command writes a json file to the specified file with the
milestone and issue summary across the provided repositories. The output
is a map, with several layers. The outer layer is keyed on milestone
title. The layer below that is keyed on repository name. Inside each
repository is a liting of the individual open issues, number of open
issues, number of closed issues, and the milestone number.